Fix some potentially error prone code constructs

+++ src/transcode.cc
@@ -183,7 +183,7 @@ static int transcode_finish(Cache_Entry* cache_entry, FFmpeg_Transcoder &transco
                        format_size_ex(cache_entry->m_cache_info.m_predicted_filesize).c_str(),
                        format_size_ex(cache_entry->m_cache_info.m_encoded_filesize).c_str(),
                        format_result_size_ex(cache_entry->m_cache_info.m_encoded_filesize, cache_entry->m_cache_info.m_predicted_filesize).c_str(),
-                       static_cast<double>((cache_entry->m_cache_info.m_encoded_filesize * 1000 / (cache_entry->m_cache_info.m_predicted_filesize + 1)) + 5) / 10);
+                       ((static_cast<double>(cache_entry->m_cache_info.m_encoded_filesize) * 1000 / (static_cast<double>(cache_entry->m_cache_info.m_predicted_filesize) + 1)) + 5) / 10);
     }
 
     cache_entry->flush();
